[quant][core] Refactor implementations for reference module
Summary:
* Refactored some methods to ReferenceQuantizedModule
** _init_weight_qparams, get_quantized_weight, get_wegiht
* Added float16 conversion in _quantize_weight

Test Plan:
python test/test_quantization.py TestQuantizeFx
python test/test_quantization.py TestQuantizeFxOps

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
